Specify UTF-8 as encoding when reading source file
Relying on the default charset should be avoided as it varies based on
locale/collation. Setting it to UTF-8 should be generally safe, as files
saved with ASCII or ISO-8859-1 encoding will still open correctly.

Without it, an environment such as the Google cloud-builer for sbt will
error out if the file is saved as UTF-8.